Invoice details - linked submissions tab - proof of concept - iteration 1
It has been identified that there is a user need to be able to quickly access historical invoice details as they were recorded for each specific submission. To achieve this we are initially investigating adding a new tab to the existing invoice details screen, which will list all the submissions that are associated with that invoice.